requirements.txt
customtkinter
tkcolorpicker
hidapi
pyaudiowpatch
pynput
numpy
pillow
pystray